Some small cleanups: * add a missing copyright header * change an include guard to reflect the filename * use qreal instead of float
svn path=/trunk/KDE/kdebase/workspace/; revision=832313

+++ plasma/dataengines/nowplaying/amarok.cpp
@@ -197,7 +197,7 @@ void Amarok::next()
     }
 }
 
-void Amarok::setVolume(float volume) {
+void Amarok::setVolume(qreal volume) {
     if (amarokPlayer->isValid()) {
         amarokPlayer->setVolume(volume * 100);
     }
